fix: button to resume assigned errands in casedata
