+++ frontend/src/casedata/services/casedata-errand-service.ts
@@ -819,7 +819,7 @@ export const isAdmin: (errand: IErrand, user: User) => boolean = (errand, user) 
   return user.username.toLocaleLowerCase() === errand?.administrator?.adAccount?.toLocaleLowerCase();
 };
 
-export const setSuspendedErrands = async (
+export const setErrandStatus = async (
   errandId: number,
   municipalityId: string,
   status: ErrandStatus,
